+++ example/lib/pages/grids_page.dart
@@ -41,7 +41,7 @@ class _GridPage extends State<GridPage> {
             ResponsiveContainer(
 
               // Determines the container's limit size
-              maxWidth: ResponsiveScreen.limits['lg'],
+              maxWidth: ResponsiveScreen.limits[ScreenSize.lg],
 
               margin: EdgeInsets.symmetric(horizontal: 10),
 
@@ -62,78 +62,78 @@ class _GridPage extends State<GridPage> {
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
                           gridSizes: {
-                            'xs': 12
+                            ScreenSize.xs: 12
                           },
-                          backgroundColor: ResponsiveScreen.isScreenSize('us', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.us, screenSize) ? Colors.green : Colors.red,
                           children: [
-                            Text('Is Ultra Small Screen:\n('+ResponsiveScreen.limits['us'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('us', screenSize) ? 'true' : 'false')
+                            Text('Is Ultra Small Screen:\n('+ResponsiveScreen.limits[ScreenSize.us].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.us, screenSize) ? 'true' : 'false')
                           ]
                       ),
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
                           gridSizes: {
-                            'xs': 12
+                            ScreenSize.xs: 12
                           },
-                          backgroundColor: ResponsiveScreen.isScreenSize('xs', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.xs, screenSize) ? Colors.green : Colors.red,
                           children: [
-                            Text('Is Extra Small Screen:\n('+ResponsiveScreen.limits['xs'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('xs', screenSize) ? 'true' : 'false')
+                            Text('Is Extra Small Screen:\n('+ResponsiveScreen.limits[ScreenSize.xs].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.xs, screenSize) ? 'true' : 'false')
                           ]
                       ),
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
-                          backgroundColor: ResponsiveScreen.isScreenSize('sm', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.sm, screenSize) ? Colors.green : Colors.red,
                           gridSizes: {
-                            'xs': 12
+                            ScreenSize.xs: 12
                           },
                           children: [
-                            Text('Is Small Screen:\n('+ResponsiveScreen.limits['sm'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('sm', screenSize) ? 'true' : 'false')
+                            Text('Is Small Screen:\n('+ResponsiveScreen.limits[ScreenSize.sm].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.sm, screenSize) ? 'true' : 'false')
                           ]
                       ),
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
-                          backgroundColor: ResponsiveScreen.isScreenSize('md', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.md, screenSize) ? Colors.green : Colors.red,
                           gridSizes: {
-                            'xs': 12
+                            ScreenSize.xs: 12
                           },
                           children: [
-                            Text('Is Medium Screen:\n('+ResponsiveScreen.limits['md'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('md', screenSize) ? 'true' : 'false')
+                            Text('Is Medium Screen:\n('+ResponsiveScreen.limits[ScreenSize.md].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.md, screenSize) ? 'true' : 'false')
                           ]
                       ),
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
-                          backgroundColor: ResponsiveScreen.isScreenSize('lg', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.lg, screenSize) ? Colors.green : Colors.red,
                           gridSizes: {
-                            'xs': 12
+                            ScreenSize.xs: 12
                           },
                           children: [
-                            Text('Is Large Screen:\n('+ResponsiveScreen.limits['lg'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('lg', screenSize) ? 'true' : 'false')
+                            Text('Is Large Screen:\n('+ResponsiveScreen.limits[ScreenSize.lg].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.lg, screenSize) ? 'true' : 'false')
                           ]
                       ),
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
-                          backgroundColor: ResponsiveScreen.isScreenSize('xl', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.xl, screenSize) ? Colors.green : Colors.red,
                           gridSizes: {
-                            'xs': 12
+                            ScreenSize.xs: 12
                           },
                           children: [
-                            Text('Is Extra Large Screen:\n('+ResponsiveScreen.limits['xl'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('xl', screenSize) ? 'true' : 'false')
+                            Text('Is Extra Large Screen:\n('+ResponsiveScreen.limits[ScreenSize.xl].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.xl, screenSize) ? 'true' : 'false')
                           ]
                       ),
                       ResponsiveCol(
                           padding: EdgeInsets.all(10),
-                          backgroundColor: ResponsiveScreen.isScreenSize('ul', screenSize) ? Colors.green : Colors.red,
+                          backgroundColor: ResponsiveScreen.isScreenSize(ScreenSize.ul, screenSize) ? Colors.green : Colors.red,
                           gridSizes: {
-                            'ul': 12
+                            ScreenSize.ul: 12
                           },
                           children: [
-                            Text('Is Ultra Large Screen:\n('+ResponsiveScreen.limits['ul'].toString()+'px)', style: JSXTypography.h4),
-                            Text( ResponsiveScreen.isScreenSize('ul', screenSize) ? 'true' : 'false')
+                            Text('Is Ultra Large Screen:\n('+ResponsiveScreen.limits[ScreenSize.ul].toString()+'px)', style: JSXTypography.h4),
+                            Text( ResponsiveScreen.isScreenSize(ScreenSize.ul, screenSize) ? 'true' : 'false')
                           ]
                       ),
 
@@ -160,7 +160,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.orange,
-                        gridSizes: { 'xs' : 4 },
+                        gridSizes: { ScreenSize.xs : 4 },
                         children: [
                           Text('XS-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -169,7 +169,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'xs' : 4 },
+                        gridSizes: { ScreenSize.xs : 4 },
                         children: [
                           Text('XS-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -178,7 +178,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.blue,
-                        gridSizes: { 'xs' : 4 },
+                        gridSizes: { ScreenSize.xs : 4 },
                         children: [
                           Text('XS-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -189,7 +189,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.orange,
-                        gridSizes: { 'sm' : 4 },
+                        gridSizes: { ScreenSize.sm : 4 },
                         children: [
                           Text('SM-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -198,7 +198,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'sm' : 4 },
+                        gridSizes: { ScreenSize.sm : 4 },
                         children: [
                           Text('SM-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -207,7 +207,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.blue,
-                        gridSizes: { 'sm' : 4 },
+                        gridSizes: { ScreenSize.sm : 4 },
                         children: [
                           Text('SM-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -218,7 +218,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.orange,
-                        gridSizes: { 'sm' : 4 },
+                        gridSizes: { ScreenSize.md : 4 },
                         children: [
                           Text('MD-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -227,7 +227,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'sm' : 4 },
+                        gridSizes: { ScreenSize.md : 4 },
                         children: [
                           Text('MD-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -236,7 +236,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.blue,
-                        gridSizes: { 'sm' : 4 },
+                        gridSizes: { ScreenSize.md : 4 },
                         children: [
                           Text('MD-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -247,7 +247,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.orange,
-                        gridSizes: { 'lg' : 4 },
+                        gridSizes: { ScreenSize.lg : 4 },
                         children: [
                           Text('LG-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -256,7 +256,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'lg' : 4 },
+                        gridSizes: { ScreenSize.lg : 4 },
                         children: [
                           Text('LG-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -265,7 +265,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.blue,
-                        gridSizes: { 'lg' : 4 },
+                        gridSizes: { ScreenSize.lg : 4 },
                         children: [
                           Text('LG-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -276,7 +276,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.orange,
-                        gridSizes: { 'xl' : 4 },
+                        gridSizes: { ScreenSize.xl : 4 },
                         children: [
                           Text('XL-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -285,7 +285,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'xl' : 4 },
+                        gridSizes: { ScreenSize.xl : 4 },
                         children: [
                           Text('XL-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -294,7 +294,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.blue,
-                        gridSizes: { 'xl' : 4 },
+                        gridSizes: { ScreenSize.xl : 4 },
                         children: [
                           Text('XL-4', style: JSXTypography.h4),
                           Text('\n\n')
@@ -318,14 +318,14 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'xs' : 4 },
-                        gridOffsetSizes: { 'xs' : 4 },
+                        gridSizes: { ScreenSize.xs : 4 },
+                        gridOffsetSizes: { ScreenSize.xs : 4 },
                         children: _aLotOfText
                     ),
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.blue,
-                        gridSizes: { 'xs' : 4 },
+                        gridSizes: { ScreenSize.xs : 4 },
                         children: _aLotOfText
                     ),
 
@@ -340,7 +340,7 @@ class _GridPage extends State<GridPage> {
                     ResponsiveCol(
                         padding: EdgeInsets.all(10),
                         backgroundColor: Colors.green,
-                        gridSizes: { 'xs' : 12 },
+                        gridSizes: { ScreenSize.xs : 12 },
                         children: _smallText
                     ),
 
@@ -350,10 +350,10 @@ class _GridPage extends State<GridPage> {
                           padding: EdgeInsets.all(10),
                           backgroundColor: Colors.blue,
                           gridSizes: {
-                            'xs' : 4,
-                            'sm' : 3,
-                            'lg' : 2,
-                            'xl' : 1,
+                            ScreenSize.xs : 4,
+                            ScreenSize.sm : 3,
+                            ScreenSize.lg : 2,
+                            ScreenSize.xl : 1,
                           },
                           children: _smallText
                       )