+++ example/lib/pages/home_page.dart
@@ -27,10 +27,10 @@ class _HomePage extends State<HomePage> {
               padding: EdgeInsets.all(10),
               backgroundColor: Colors.blue,
               gridSizes: {
-                'xs' : 4,
-                'sm' : 3,
-                'lg' : 2,
-                'xl' : 1,
+                ScreenSize.xs : 4,
+                ScreenSize.sm : 3,
+                ScreenSize.lg : 2,
+                ScreenSize.xl : 1,
               },
               children: [
                 Text('Lorem ipsum dolor sit amet, consectetur adipiscing elit')
@@ -38,75 +38,75 @@ class _HomePage extends State<HomePage> {
           )
       ).toList();
 
-    MediaQueryData mediaQuery = MediaQuery.of(context);
+    ThemeData themeData = Theme.of(context);
+    MediaQueryData mediaQueryData = MediaQuery.of(context);
 
     return Scaffold(
         appBar: AppBar(
           title: const Text('Home', overflow: TextOverflow.ellipsis),
         ),
         drawer: Sidebar(),
-        body: Container(
+        body: Container( //background
+          width: mediaQueryData.size.width,
+          height: mediaQueryData.size.height,
           color: Color(0xFFCCCCCC),
-          child: ListView(
-            children: <Widget>[
-              ResponsiveContainer(
-                margin: EdgeInsets.symmetric(vertical: 10),
-                padding: EdgeInsets.symmetric(horizontal: 10),
-                backgroundColor: Colors.white,
-                maxWidth: mediaQuery.size.width * 0.95,
-                children: <Widget>[
-                  ResponsiveRow(
-                    margin: EdgeInsets.symmetric(vertical: 10),
-                    children: <Widget>[
+          child: SingleChildScrollView(
+            child: ResponsiveContainer(
+              padding: EdgeInsets.symmetric(horizontal: 10),
+              backgroundColor: Colors.white,
+              maxWidth: ResponsiveScreen.limits[ScreenSize.lg],//mediaQuery.size.width * 0.95,
+              children: <Widget>[
+                ResponsiveRow(
+                  margin: EdgeInsets.symmetric(vertical: 10),
+                  children: <Widget>[
 
-                      ResponsiveCol(
-                          padding: EdgeInsets.all(10),
-                          margin: EdgeInsets.only(bottom: 20),
-                          backgroundColor: Colors.blueGrey,
-                          children: [
-                            Text('Flutter Responsive Layout', style: JSXTypography.h4.merge(TextStyle(color: Colors.white)))
-                          ]
-                      ),
-                    ]
-                  ),
-                  ResponsiveRow(
-                    margin: EdgeInsets.symmetric(vertical: 10),
-                    children: <Widget>[
-
-                      // By default, the column occupies the entire row, always
-                      ResponsiveCol(
+                    ResponsiveCol(
+                        padding: EdgeInsets.all(10),
+                        margin: EdgeInsets.only(bottom: 20),
+                        backgroundColor: Colors.blueGrey,
                         children: [
-                          JSX(
-                            '<div>'
-                                '<h3>Responsive Layouts</h3><h6>for <i>Flutter</i></h6>'
-                                '<br><br>'
-                                '<p>This <b>RichText</b> was easily produced and personalized using pure HTML</p>'
-                                '<p>Bellow there is an example of 12 columns, wich changes the amount of each line depending of his father´s widget size.</p>'
-                            '</div>',
-                            display: DisplayLine.block,
-                            padding: EdgeInsets.only(bottom: 20),
-                            stylesheet: {
-                              'h3': JSXStylesheet(
-                                  textStyle: TextStyle(color: Theme.of(context).primaryColor),
-                                  displayLine: DisplayLine.block
-                              ),
-                              'h6': JSXStylesheet(
-                                  textStyle: TextStyle(color: Theme.of(context).primaryColor),
-                                  displayLine: DisplayLine.block
-                              )
-                            },
-                            margin: EdgeInsets.symmetric(horizontal: 10, vertical: 20),
-                          )
+                          Text('Flutter Responsive Layout', style: JSXTypography.h4.merge(TextStyle(color: Colors.white)))
                         ]
-                      )
+                    ),
+                  ]
+                ),
+                ResponsiveRow(
+                  margin: EdgeInsets.symmetric(vertical: 10),
+                  children: <Widget>[
 
-                    ]..addAll(
-                        responsiveGridExampe
+                    // By default, the column occupies the entire row, always
+                    ResponsiveCol(
+                      children: [
+                        JSX(
+                          '<div>'
+                              '<h3>Responsive Layouts</h3>'
+                              '<h6>for <i>Flutter</i></h6>'
+                              '<br><br>'
+                              '<p>Bellow there is an example of 12 columns, which changes the amount of each line depending of his father\'s widget size.</p>'
+                          '</div>',
+                          display: DisplayLine.block,
+                          padding: EdgeInsets.only(bottom: 20),
+                          stylesheet: {
+                            'h3': JSXStylesheet(
+                                textStyle: TextStyle(color: themeData.primaryColor),
+                                displayLine: DisplayLine.block
+                            ),
+                            'h6': JSXStylesheet(
+                                textStyle: TextStyle(color: themeData.primaryColor),
+                                displayLine: DisplayLine.block
+                            )
+                          },
+                          margin: EdgeInsets.symmetric(horizontal: 10, vertical: 20),
+                        )
+                      ]
                     )
+
+                  ]..addAll(
+                      responsiveGridExampe
                   )
-                ],
-              )
-            ],
+                )
+              ],
+            )
           ),
         )
     );