+++ lib/flutter_responsive.dart
@@ -1,5 +1,6 @@
 library flutter_responsive;
 
+export 'src/responsive_enum.dart';
 export 'src/responsive_row.dart';
 export 'src/responsive_col.dart';
 export 'src/responsive_screen.dart';