+++ lib/src/responsive_enum.dart
@@ -1,0 +1,22 @@
+enum ScreenSize {
+  /// Smart watches (0.00 >= width > 310.00)
+  us,
+
+  /// Small phones (5S) (310.00 >= width > 576.00)
+  xs,
+
+  /// Medium phones (576.00 >= width > 768.00)
+  sm,
+
+  /// Large phones (iPhone X) (768.00 >= width > 992.00)
+  md,
+
+  /// Tablets (992.00 >= width > 1200.00)
+  lg,
+
+  /// Laptops (1200.00 >= width > 2000.00)
+  xl,
+
+  /// TVs and 4k ( width > 2000.00)
+  ul,
+}